fix: address code review — dashboard categories_summary, schema defaults, and code style
- dashboard_service.py: categories_summary now uses PAID installments
  instead of actual_amount (prevents inflated category spending)
- schemas.py: SupplierPatchIn.website default changed from '' to None
- schemas.py: SupplierOut.state uses min/max length instead of regex
- tests: merged duplicate BudgetCategoryTotalSpent tests
- tests: updated dashboard percentage assertion from 40 to 20
- supplierFormSchema.ts: import { z } from zod (convention)
- SupplierFormDialog.tsx: explicit useForm type param
- Contract sync: openapi.json and Orval hooks regenerated
